Revert "Refactored VAO creation code"
This reverts commit 23c2be82

+++ RGBEngine/src/main/java/extras/tiles/Tilemap.java
@@ -100,8 +100,7 @@ public class Tilemap extends XMLElement {
                             throw new IllegalStateException("Data is badly formatted");
                         }
                         for (int x = 0; x < width; x++) {
-
-                            tiles[x][y] = (int) Long.parseLong(row[x]);
+                            tiles[x][y] = Integer.parseInt(row[x]);
                         }
                     }
                 } else {
